Add files via upload
To achieve the recognition of multi-attribute of object, I redesign the mnist dataset, change the color, size, location of the number. Meanwhile, I change the label accordingly. The deep neural network I use is the most common convolution neural network. Through test,we can conclude that we can use one neural network to recognize multi-attribute so long as the attribute difference of objects can be represented by functions. The concrete network(generation network) can generate the output which the input rarely contained from the attributes the network learned. Its generalization ability is good because the network is a continuous function. Through one more test, We can conclude that one neural network can do image recognition, speech recognition,and nature language processing and other things so long as the output node and the input node and more parameters add into the network. The network is universal so long as the network can process different inputs. I guess that the phenomenon of synaesthesia is the result of multi-input and multi-output. I guess that connection in mind can realize through the universal network and sending the output into input.